+++ src/annbatch/samplers/__init__.py
@@ -1,4 +1,4 @@
-from ._categorical_sampler import ClassSampler
+from ._class_sampler import ClassSampler
 from ._distributed_sampler import DistributedSampler
 from ._random_sampler import RandomSampler
 from ._sequential_sampler import SequentialSampler